+++ packages/components/src/Center/Center.test.tsx
@@ -1,5 +1,4 @@
 import { render, screen } from '@testing-library/react';
-import React from 'react';
 import { Center } from './Center';
 
 test('supports classnames per default and maxWidth prop', () => {