+++ packages/components/src/Container/Container.test.tsx
@@ -1,5 +1,4 @@
 import { screen } from '@testing-library/react';
-import React from 'react';
 import { Theme, cva } from '@marigold/system';
 import { Headline } from '../Headline/Headline';
 import { Text } from '../Text/Text';